Fix routes and types, remove etags
