+++ app/routes.ts
@@ -7,15 +7,15 @@ import {
 } from '@react-router/dev/routes';
 
 export default [
-  layout('./routes/layout.tsx', [
-    index('./routes/home.tsx'),
-    route('404', './routes/404.tsx'),
+  layout('./routes/_news/route.tsx', [
+    index('./routes/_news.index/route.tsx'),
+    route('404', './routes/_news.404/route.tsx'),
   ]),
+  route('news/:name', './routes/news.$name/route.tsx'),
   ...prefix('docs', [
-    index('./routes/docs.index.tsx'),
-    route(':name', './routes/docs.$name.tsx'),
+    index('./routes/docs.index/route.tsx'),
+    route(':name', './routes/docs.$name/route.tsx'),
   ]),
-  route('news/:name', './routes/news.$name.tsx'),
-  route('theme', './routes/theme.tsx'),
-  route('*', './routes/$.tsx'),
+  route('theme', './routes/theme/route.tsx'),
+  route('*', './routes/$/route.tsx'),
 ] satisfies RouteConfig;
