+++ app/routes/_news/route.tsx
@@ -3,9 +3,9 @@ import { getGitFeed } from '~/.server/get-git-feed';
 import { getLatestRelease } from '~/.server/get-latest-release';
 import { GitFeed } from '~/components/git-feed';
 import { Release } from '~/components/release';
-import type { Route } from './+types/layout';
+import type { Route } from './+types/route';
 
-export async function loader({ request }: Route.LoaderArgs) {
+export async function loader() {
   const commits = await getGitFeed();
   const release = await getLatestRelease();
 
@@ -17,7 +17,7 @@ export default function Component({ loaderData }: Route.ComponentProps) {
 
   return (
     <div className="grid gap-12 lg:grid-cols-12 lg:gap-16">
-      <div className="space-y-6 lg:col-span-7">
+      <div className="space-y-6 lg:col-span-8">
         <header>
           <h1 className="mb-1 font-bold text-2xl text-amber-12">Latest News</h1>
           <hr className="border-sand-8" />
@@ -25,8 +25,8 @@ export default function Component({ loaderData }: Route.ComponentProps) {
         <Outlet />
       </div>
       <hr className="border-sand-8 lg:hidden" />
-      <div className="grid gap-4 self-start lg:col-span-5">
-        {release ? <Release release={release} /> : null}
+      <div className="grid gap-4 self-start lg:col-span-4">
+        <Release release={release} />
         <GitFeed commits={commits} />
       </div>
     </div>
